+++ package.json
@@ -20,7 +20,6 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^8.4.1",
-    "@types/jest": "^29.5.12",
     "@types/node": "^20.12.2",
     "@types/pg": "^8.11.4",
     "@vitest/coverage-v8": "^1.6.0",
