+++ src/database.ts
@@ -1,6 +1,6 @@
 import { Pool } from 'pg';
 import { drizzle } from 'drizzle-orm/node-postgres';
-import { schema } from './schemas';
+import { schema } from './schemas/index';
 
 export const pool = new Pool({
   host: 'localhost',
